+++ src/ethereum/EthereumEngine.ts
@@ -1164,24 +1164,13 @@ export class EthereumEngine extends CurrencyEngine<
       // Use an unconfirmed nonce if
       // 1. We have unconfirmed spending txs in the transaction list
       // 2. It is greater than the confirmed nonce
-      // 3. Is no more than 5 higher than confirmed nonce
       // Otherwise, use the next nonce
       if (
         this.walletLocalData.numUnconfirmedSpendTxs != null &&
         gt(this.otherData.unconfirmedNextNonce, this.otherData.nextNonce)
       ) {
-        const diff = sub(
-          this.otherData.unconfirmedNextNonce,
-          this.otherData.nextNonce
-        )
-        if (lte(diff, '5')) {
-          nonce = this.otherData.unconfirmedNextNonce
-          this.walletLocalDataDirty = true
-        } else {
-          const e = new Error('Excessive pending spend transactions')
-          e.name = 'ErrorExcessivePendingSpends'
-          throw e
-        }
+        nonce = this.otherData.unconfirmedNextNonce
+        this.walletLocalDataDirty = true
       } else {
         nonce = this.otherData.nextNonce
       }